+++ spec/requests/federation/activities_spec.rb
@@ -27,6 +27,17 @@ RSpec.describe '/federation/activities', type: :request do
       expect(response).to be_successful
     end
 
+    it 'rejects badly-signed requests' do
+      get fedipub.server_actor_outbox_url(local_actor), headers: { accept: Mime[:activitypub], signature: 'poop' }
+      expect(response).to have_http_status :unauthorized
+    end
+
+    it 'rejects unsigned requests when signatures are required' do
+      allow(Fedipub::ServerController).to receive(:require_signature?).and_return(true)
+      get fedipub.server_actor_outbox_url(local_actor), headers: { accept: Mime[:activitypub] }
+      expect(response).to have_http_status :unauthorized
+    end
+
     ACTIVITYPUB_CONTENT_TYPES.each do |accept|
       it "responds with LD in response to a #{accept} request" do
         get fedipub.server_actor_outbox_url(local_actor), headers: { accept: accept }
@@ -111,6 +122,17 @@ RSpec.describe '/federation/activities', type: :request do
     it 'renders a successful response' do
       get fedipub.server_actor_activity_url(activity.actor.to_param, activity.to_param), headers: { accept: Mime[:activitypub] }
       expect(response).to be_successful
+    end
+
+    it 'rejects badly-signed requests' do
+      get fedipub.server_actor_activity_url(activity.actor.to_param, activity.to_param), headers: { accept: Mime[:activitypub], signature: 'poop' }
+      expect(response).to have_http_status :unauthorized
+    end
+
+    it 'rejects unsigned requests when signatures are required' do
+      allow(Fedipub::ServerController).to receive(:require_signature?).and_return(true)
+      get fedipub.server_actor_activity_url(activity.actor.to_param, activity.to_param), headers: { accept: Mime[:activitypub] }
+      expect(response).to have_http_status :unauthorized
     end
 
     ACTIVITYPUB_CONTENT_TYPES.each do |accept|